Learned how to use the this keyword...sort of
If an object literal is a tool bag, the set_key function is a little robot whose job it is to rummage around inside the toolbag, find the hammer, throw the hammer out of the bag, build a new, slightly different hammer, and pretend it was the same hammer.  The problem is that the robot is a bit dim.  It doesn't know that you want the hammer inside the toolbag. So you have to tell it, "In 'this' toolbag, go replace the hammer." If you don't specifically tell it that you want it to be in this toolbag, it will go out into the world and look for a hammer to replace, and replace that hammer instead.