Learned how to construct a javascript object using a constructor
A javascript object is a toolbag.  A constructor is the procedure for putting tools in the toolbag. After all, different constructon workers need different tools for different projects, and will put them in different bags.

When you say my toolbag = new Toolbag('hammer', 'wrench'), it's like handing a bunch of tools to the new guy in the shop person and telling them to make you a bag of the tools.