Added a new function to a javascript object using the prototype method.
A javascript object is like a toolbag.  The properties are like tools. Objects can also have power tools, which are properties that are set to be functions.

However, what if you make a bag of tools.  Then, later, you decide that you really want a power tool in the bag?  You can add the power tool after you've alreay put together the bag. You do this with the protoype method.